backend/project/requirements.txt
# Django core
Django>=5.2.1,<6.0

# Django REST Framework
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0

# CORS headers
django-cors-headers>=4.3.0,<5.0

# Environment variables
python-decouple>=3.8,<4.0

# Twilio for SMS
twilio>=8.10.0,<9.0

# Channels for WebSocket support
channels>=4.0.0,<5.0
channels-redis>=4.1.0,<5.0

# ASGI server
daphne>=4.0.0,<5.0

# PostgreSQL adapter (for production)
psycopg2-binary>=2.9.9,<3.0

# Production WSGI server (alternative to daphne for HTTP)
gunicorn>=21.2.0,<22.0


drf-spectacular>=0.27.0,<1.0